requirements.txt
pandas==1.3.3
numpy==1.21.2
requests==2.26.0
psycopg2-binary==2.9.1
